rock rest added and changed the rating to double

+++ Transpire/app/src/main/java/com/brodyclark_l2/transpire/data/Location.kt
@@ -4,4 +4,4 @@ package com.brodyclark_l2.transpire.data
 import java.io.Serializable
 import java.util.*
 
-data class location(val name:String, val addressLink:String, val lowAge:Int, val highAge:Int, val rating: Int, var id: UUID = UUID.randomUUID()): Serializable
+data class location(val name:String, val addressLink:String, val lowAge:Int, val highAge:Int, val rating: Double, var id: UUID = UUID.randomUUID()): Serializable
